requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.91
botocore==1.34.91
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
ecdsa==0.19.0
jmespath==1.0.1
kombu==5.3.7
prompt-toolkit==3.0.43
pyasn1==0.6.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.4
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
